Prototype Version of the Account Details interface
Still untested, has no current links between any of the main php pages.  Still needs work on grabbing the DB connection from the config file.  HTML sections can be debugged, but there are still questions as to take the results of the array (from the DB) and stylize it with HTML.